+++ tests/Feature/BaseUrlModeTest.php
@@ -3,6 +3,7 @@
 namespace Estin92\DvlaVes\Tests\Feature;
 
 use Estin92\DvlaVes\Contracts\VehicleEnquiry;
+use Estin92\DvlaVes\Exceptions\DvlaVesException;
 use Estin92\DvlaVes\Services\VehicleEnquiryService;
 use Estin92\DvlaVes\Tests\TestCase;
 use Illuminate\Support\Facades\Http;
@@ -68,5 +69,32 @@ class BaseUrlModeTest extends TestCase
         $this->makeService()->lookup('AB12CDE');
 
         $this->assertCalledHost('https://new-dvla-host.example');
+    }
+
+    public function test_a_plain_http_override_is_upgraded_to_https(): void
+    {
+        config([
+            'dvla-ves.mode' => 'prod',
+            'dvla-ves.base_urls.prod' => 'http://new-dvla-host.example',
+        ]);
+
+        Http::fake(['*' => Http::response(['registrationNumber' => 'AB12CDE'], 200)]);
+
+        $this->makeService()->lookup('AB12CDE');
+
+        $this->assertCalledHost('https://new-dvla-host.example');
+    }
+
+    public function test_a_non_http_base_url_is_rejected(): void
+    {
+        config([
+            'dvla-ves.mode' => 'prod',
+            'dvla-ves.base_urls.prod' => 'ftp://new-dvla-host.example',
+        ]);
+
+        $this->expectException(DvlaVesException::class);
+        $this->expectExceptionMessage('must use HTTPS');
+
+        $this->makeService();
     }
 }